updated base.py. (winfunc added in logfbank)
Previously I was using fbank for feature extraction. Then when I found that when I used log on the Fbank the accuracy of the model was better. But I was using hamming frame on windows for feature extraction, because of this I was taking log after implementing fbank. So it would be better if we have a choice to apply window function while using logfbank function as well.